Fixes pointer handling to allow scrolling
The Pointer Interface of IE uses a new way to handle
certain actions by the user scripts while leaving other
actions for the browser. For the Touch Interface we already
implemented a threshold to signal back to the UA if we want
to allow it to scroll if the user is not swiping.